+++ app/application/forms/errors/date_request.yml
@@ -1,3 +1,3 @@
 en:
   errors:
-    date: 'Invalid date. Please follow mm/dd/yyyy'
+    game_date: 'Invalid date. Please follow mm/dd/yyyy'
